fix(messaging): address review findings

+++ src/mediator-framework/Ark.Tools.MediatorFramework.Messaging/InMemoryMessagingTransport.cs
@@ -84,7 +84,7 @@ public sealed class InMemoryMessagingTransport :
         ArgumentNullException.ThrowIfNull(headers);
         return headers.Sum(static pair =>
             (long)System.Text.Encoding.UTF8.GetByteCount(pair.Key)
-            + System.Text.Encoding.UTF8.GetByteCount(pair.Value));
+            + (long)System.Text.Encoding.UTF8.GetByteCount(pair.Value));
     }
 
     /// <summary>Configures the native retry limit and delay for a queue.</summary>
